Componentized AnimatedNumbers to separate it from the Filters component
